+++ src/Ar/ATN/Director.cpp
@@ -13,7 +13,7 @@
 
 using namespace atn;
 
-Director::Director(/* args */)
+Director::Director(/* args */) : outstream(0)
 {
 }
 
@@ -22,119 +22,119 @@ Director::~Director()
 
 }
 
-void Director::addState( const std::string state,  AtnAPIState_typ* api, void *_pParameters, size_t _sParameters ){
+void Director::addState( const std::string state,  AtnAPIState_typ* api, void *_pParameters, size_t _sParameters, const std::string& taskName ){
 
     auto it = states.find(state);
 
     if (it != states.end()){
-        it->second.subscribe( api, _pParameters, _sParameters);
+        it->second.subscribe( api, _pParameters, _sParameters, taskName);
     }
     else{
         State newAction( state );
-        newAction.subscribe(api, _pParameters, _sParameters);
+        newAction.subscribe(api, _pParameters, _sParameters, taskName);
         states.insert( std::pair<std::string, State>(state, newAction));
     }
 }
 
 
-void Director::addState( const std::string state, const std::string name, char *moduleStatus, plcbit* moduleByPass, bool* value, void *_pParameters, size_t _sParameters ){
+void Director::addState( const std::string state, const std::string name, char *moduleStatus, plcbit* moduleByPass, bool* value, void *_pParameters, size_t _sParameters, const std::string& taskName ){
 
 	auto it = states.find(state);
 
 	if (it != states.end()){
-		it->second.subscribe( name, value, moduleByPass, moduleStatus, _pParameters, _sParameters);
+		it->second.subscribe( name, value, moduleByPass, moduleStatus, _pParameters, _sParameters, taskName);
 	}
 	else{
 		State newAction( state );
-		newAction.subscribe( name, value, moduleByPass, moduleStatus, _pParameters, _sParameters);
+		newAction.subscribe( name, value, moduleByPass, moduleStatus, _pParameters, _sParameters, taskName);
 		states.insert( std::pair<std::string, State>(state, newAction));
 	}
 
 
 }
 
-void Director::addStateBool( const std::string state, const std::string name, bool* value ){
+void Director::addStateBool( const std::string state, const std::string name, bool* value, const std::string& taskName ){
 
 	auto it = states.find(state);
 
 	if (it != states.end()){
-		it->second.subscribe( name, value);
+		it->second.subscribe( name, value, taskName);
 	}
 	else{
 		State newAction( state );
-		newAction.subscribe( name, value);
+		newAction.subscribe( name, value, taskName);
 		states.insert( std::pair<std::string, State>(state, newAction));
 	}
 }
 
-void Director::addResourceBool( const std::string state, const std::string name, unsigned long int *pResourceUid, bool *value ){
+void Director::addResourceBool( const std::string state, const std::string name, unsigned long int *pResourceUid, bool *value, const std::string& taskName ){
 
 	auto it = states.find(state);
 
 	if (it != states.end()){
-		it->second.subscribe( name, pResourceUid, value);
+		it->second.subscribe( name, pResourceUid, value, taskName);
 	}
 	else{
 		State newAction( state );
-		newAction.subscribe( name, pResourceUid, value);
+		newAction.subscribe( name, pResourceUid, value, taskName);
 		states.insert( std::pair<std::string, State>(state, newAction));
 	}
 }
 
 
 
-void Director::addStateBool( const std::string state, const std::string name, bool* value,  void *_pParameters, size_t _sParameters){
+void Director::addStateBool( const std::string state, const std::string name, bool* value,  void *_pParameters, size_t _sParameters, const std::string& taskName){
 
 	auto it = states.find(state);
 
 	if (it != states.end()){
-		it->second.subscribe( name, value, _pParameters, _sParameters);
+		it->second.subscribe( name, value, _pParameters, _sParameters, taskName);
 	}
 	else{
 		State newAction( state );
-		newAction.subscribe( name, value, _pParameters, _sParameters);
+		newAction.subscribe( name, value, _pParameters, _sParameters, taskName);
 		states.insert( std::pair<std::string, State>(state, newAction));
 	}
 }
 
 //Registers a bool to be automatically monitored, without full API support
-void Director::addCommandBool( const std::string command, const std::string moduleName, bool *check ){
+void Director::addCommandBool( const std::string command, const std::string moduleName, bool *check, const std::string& taskName ){
 	auto it = commands.find(command);
 
 	if (it != commands.end()){
-		it->second.subscribe( moduleName, check);
+		it->second.subscribe( moduleName, check, taskName);
 	}
 	else{
 		State newAction( command );
-		newAction.subscribe( moduleName, check);
+		newAction.subscribe( moduleName, check, taskName);
 		commands.insert( std::pair<std::string, State>(command, newAction));
 	}
 }
 
 //Registers a bool to be automatically monitored, without full API support
-void Director::addCommandPLCOpen( const std::string command, const std::string moduleName, bool * commandBit, AtnPlcOpenStatus * status ){
+void Director::addCommandPLCOpen( const std::string command, const std::string moduleName, bool * commandBit, AtnPlcOpenStatus * status, const std::string& taskName ){
 	auto it = commands.find(command);
 
 	if (it != commands.end()){
-		it->second.subscribe( moduleName, commandBit, status);
+		it->second.subscribe( moduleName, commandBit, status, taskName);
 	}
 	else{
 		State newAction( command );
-		newAction.subscribe( moduleName, commandBit, status);
+		newAction.subscribe( moduleName, commandBit, status, taskName);
 		commands.insert( std::pair<std::string, State>(command, newAction));
 	}
 }
 
 //Registers a bool to be automatically monitored, without full API support
-void Director::addCommandPLCOpen( const std::string command, const std::string moduleName, bool * commandBit, AtnPlcOpenStatus * status,  void *_pParameters, size_t _sParameters ){
+void Director::addCommandPLCOpen( const std::string command, const std::string moduleName, bool * commandBit, AtnPlcOpenStatus * status,  void *_pParameters, size_t _sParameters, const std::string& taskName ){
 	auto it = commands.find(command);
 
 	if (it != commands.end()){
-		it->second.subscribe( moduleName, commandBit, status, _pParameters, _sParameters);
+		it->second.subscribe( moduleName, commandBit, status, _pParameters, _sParameters, taskName);
 	}
 	else{
 		State newAction( command );
-		newAction.subscribe( moduleName, commandBit, status, _pParameters, _sParameters);
+		newAction.subscribe( moduleName, commandBit, status, _pParameters, _sParameters, taskName);
 		commands.insert( std::pair<std::string, State>(command, newAction));
 	}
 }
@@ -191,6 +191,80 @@ State * Director::getCommand( const std::string cmd ){
     }
 }
 
+bool Director::addValue( const std::string state, const std::string name, bool* valid, void* _pData, size_t _sData, size_t sReturn, const std::string& taskName ){
+
+	auto it = values.find(state);
+
+	if (it != values.end()){
+		if( it->second.count() > 0 ){
+			if( this->outstream ){
+				*this->outstream << "ATN value already registered: " << state << "\n";
+			}
+			return false;
+		}
+		it->second.subscribe( name, valid, _pData, _sData, taskName );
+		it->second.sReturn = sReturn;
+		it->second.returnTopic = (sReturn > 0) ? (state + ATN_RETURN_TOPIC_SUFFIX) : "";
+	}
+	else{
+		State newValue( state );
+		newValue.subscribe( name, valid, _pData, _sData, taskName );
+		newValue.sReturn = sReturn;
+		newValue.returnTopic = (sReturn > 0) ? (state + ATN_RETURN_TOPIC_SUFFIX) : "";
+		values.insert( std::pair<std::string, State>(state, newValue) );
+	}
+	return true;
+}
+
+State * Director::getValue( const std::string state ){
+	auto it = values.find(state);
+
+	if (it != values.end()){
+		return &it->second;
+	}
+	else{
+		return 0;
+	}
+}
+
+unsigned int Director::removeRegistration( const std::string& name, const std::string& taskName ){
+	unsigned int removed = 0;
+
+	auto v = values.find(name);
+	if( v != values.end() ){
+		removed += v->second.removeTask(taskName);
+	}
+
+	auto s = states.find(name);
+	if( s != states.end() ){
+		removed += s->second.removeTask(taskName);
+	}
+
+	auto c = commands.find(name);
+	if( c != commands.end() ){
+		removed += c->second.removeTask(taskName);
+	}
+
+	return removed;
+}
+
+unsigned int Director::removeAllForTask( const std::string& taskName ){
+	unsigned int removed = 0;
+
+	for( auto &kv : values ){
+		removed += kv.second.removeTask(taskName);
+	}
+
+	for( auto &kv : states ){
+		removed += kv.second.removeTask(taskName);
+	}
+	for( auto &kv : commands ){
+		removed += kv.second.removeTask(taskName);
+	}
+
+	return removed;
+}
+
 void Director::printStates( std::ostream &out ){
     out << "\nStates:" << "\n";
     for( auto state : states ){
@@ -205,6 +279,13 @@ void Director::printCommands( std::ostream &out ){
 		out << command.first << "\n";
 	}
 
+}
+
+void Director::printValues( std::ostream &out ){
+	out << "\nValues:" << "\n";
+	for( auto value : values ){
+		out << value.first << "\n";
+	}
 }
 
 void Director::printSystemJson( std::ostream &out ){
@@ -223,6 +304,14 @@ void Director::printSystemJson( std::ostream &out ){
 	for( auto state : states ){
 		if(comma) out << ",";
 		out << "\"" <<state.first << "\"";
+		comma = 1;
+	}
+	out << "]";
+	out << ",\"Values\":[";
+	comma = 0;
+	for( auto value : values ){
+		if(comma) out << ",";
+		out << "\"" <<value.first << "\"";
 		comma = 1;
 	}
 	out << "]";