+++ src/Ar/ATN/State.cpp
@@ -23,11 +23,12 @@ State::State( std:: string name ){
 State::~State(){};
 
 
-void State::subscribe( AtnAPIState_typ* api, void *_pParameters, size_t _sParameters ){
+void State::subscribe( AtnAPIState_typ* api, void *_pParameters, size_t _sParameters, const std::string& taskName ){
 
     PLCOpen state;
     state.pParameters = _pParameters;
     state.sParameters = _sParameters;
+    state.taskName = taskName;
     if( api ){
         state.name = api->moduleName;
         state.pValue = &(api->active);
@@ -36,16 +37,18 @@ void State::subscribe( AtnAPIState_typ* api, void *_pParameters, size_t _sParame
     this->PLCOpenState.push_back( state );
 }
 
-void State::subscribe(  const std:: string ModuleName, bool* value ){
+void State::subscribe(  const std:: string ModuleName, bool* value, const std::string& taskName ){
     PLCOpen state;
     state.name = ModuleName;
+    state.taskName = taskName;
     state.pValue = value;
 
     this->PLCOpenState.push_back( state );
 }
-void State::subscribe(  const std:: string ModuleName, unsigned long int * pUid, bool* value ){
+void State::subscribe(  const std:: string ModuleName, unsigned long int * pUid, bool* value, const std::string& taskName ){
 	PLCOpen state;
 	state.name = ModuleName;
+	state.taskName = taskName;
 	state.pResourceUser = pUid;
 	state.pValue = value;
 
@@ -56,10 +59,11 @@ void State::subscribe(  const std:: string ModuleName, unsigned long int * pUid,
 	this->PLCOpenState.push_back( state );
 }
 
-void State::subscribe(  const std:: string ModuleName, bool* active, bool* bypass, char * statusString, void *_pParameters, size_t _sParameters  ){
+void State::subscribe(  const std:: string ModuleName, bool* active, bool* bypass, char * statusString, void *_pParameters, size_t _sParameters, const std::string& taskName  ){
 
 	PLCOpen state;
 	state.name = ModuleName;
+	state.taskName = taskName;
 	state.pValue = active;
 	state.pParameters = _pParameters;
 	state.sParameters = _sParameters;
@@ -69,10 +73,11 @@ void State::subscribe(  const std:: string ModuleName, bool* active, bool* bypas
 	this->PLCOpenState.push_back( state );
 }
 
-void State::subscribe(  const std:: string ModuleName, bool* value, void *_pParameters, size_t _sParameters ){
+void State::subscribe(  const std:: string ModuleName, bool* value, void *_pParameters, size_t _sParameters, const std::string& taskName ){
 
     PLCOpen state;
     state.name = ModuleName;
+    state.taskName = taskName;
     state.pValue = value;
     state.pParameters = _pParameters;
     state.sParameters = _sParameters;
@@ -80,10 +85,11 @@ void State::subscribe(  const std:: string ModuleName, bool* value, void *_pPara
     this->PLCOpenState.push_back( state );
 }
 
-void State::subscribe(  const std:: string ModuleName,  plcbit* command, AtnPlcOpenStatus *status ){
+void State::subscribe(  const std:: string ModuleName,  plcbit* command, AtnPlcOpenStatus *status, const std::string& taskName ){
 
 	PLCOpen state;
 	state.name = ModuleName;
+	state.taskName = taskName;
 	state.pValue = command;
 	state.pStatus = &(status->status);
 	state.pActiveCommand =  (char*)&(status->activeCommand);
@@ -95,10 +101,11 @@ void State::subscribe(  const std:: string ModuleName,  plcbit* command, AtnPlcO
 	this->PLCOpenState.push_back( state );
 }
 
-void State::subscribe(  const std:: string ModuleName,  plcbit* command, AtnPlcOpenStatus *status, void *_pParameters, size_t _sParameters   ){
+void State::subscribe(  const std:: string ModuleName,  plcbit* command, AtnPlcOpenStatus *status, void *_pParameters, size_t _sParameters, const std::string& taskName   ){
 
     PLCOpen state;
     state.name = ModuleName;
+    state.taskName = taskName;
     state.pValue = command;
     state.pStatus = &(status->status);
 	state.pActiveCommand =  (char*)&(status->activeCommand);
@@ -289,6 +296,20 @@ unsigned int State::count(){
     return this->PLCOpenState.size();
 }
 
+unsigned int State::removeTask( const std::string& taskName ){
+    unsigned int removed = 0;
+    for( auto it = this->PLCOpenState.begin(); it != this->PLCOpenState.end(); ){
+        if( it->taskName == taskName ){
+            it = this->PLCOpenState.erase( it );
+            removed++;
+        }
+        else{
+            ++it;
+        }
+    }
+    return removed;
+}
+
 void State::print( std::ostream &outbuf){
     outbuf << "\nState Check: " << this->name << "\n";
     for( auto state : this->PLCOpenState ){
@@ -311,5 +332,23 @@ void State::plcopenReport( std::ostream &outbuf){
 		}
 		outbuf << "\n";
 	}
-	
+
+}
+
+void State::printValue( std::ostream &outbuf){
+	outbuf << "\nValue: " << this->name << "\n";
+	for( auto state : this->PLCOpenState ){
+		outbuf << "Owner: " << state.name;
+		if( state.pValue ){
+			outbuf << " | Valid: " << (bool) *state.pValue;
+		}
+		else{
+			outbuf << " | Valid: None";
+		}
+		outbuf << " | Size: " << state.sParameters << "\n";
+	}
+	if( !this->returnTopic.empty() ){
+		outbuf << "Return topic: " << this->returnTopic << " (" << this->sReturn << " bytes)\n";
+	}
+	outbuf << "\n";
 }