+++ src/Ar/ATN/includes/State.h
@@ -20,19 +20,21 @@ namespace atn{
 
     public:
 
-		std::vector<std::string> DependentState; 
-		std::vector<PLCOpen> PLCOpenState; 
+        std::vector<std::string> DependentState;
+        std::vector<PLCOpen> PLCOpenState;
+        std::string returnTopic;
+        size_t sReturn = 0;
 
         State( std:: string name);    
         ~State();
-        void subscribe(  AtnAPIState_typ* api, void *_pParameters, size_t _sParameters  );
-        void subscribe(  const std:: string ModuleName, bool* active);
-        void subscribe(  const std:: string ModuleName, bool* active, void *_pParameters, size_t _sParameters  );
-		void subscribe(  const std:: string ModuleName, bool* active, bool* bypass, char * statusString, void *_pParameters, size_t _sParameters  );
-		void subscribe(  const std:: string ModuleName, unsigned long int * pUid, bool* value );
-		
-		void subscribe(  const std:: string ModuleName,  bool* command, AtnPlcOpenStatus *status );
-		void subscribe(  const std:: string ModuleName,  bool* command, AtnPlcOpenStatus *status, void *_pParameters, size_t _sParameters  );
+        void subscribe(  AtnAPIState_typ* api, void *_pParameters, size_t _sParameters, const std::string& taskName  );
+        void subscribe(  const std:: string ModuleName, bool* active, const std::string& taskName);
+        void subscribe(  const std:: string ModuleName, bool* active, void *_pParameters, size_t _sParameters, const std::string& taskName  );
+		void subscribe(  const std:: string ModuleName, bool* active, bool* bypass, char * statusString, void *_pParameters, size_t _sParameters, const std::string& taskName  );
+		void subscribe(  const std:: string ModuleName, unsigned long int * pUid, bool* value, const std::string& taskName );
+
+		void subscribe(  const std:: string ModuleName,  bool* command, AtnPlcOpenStatus *status, const std::string& taskName );
+		void subscribe(  const std:: string ModuleName,  bool* command, AtnPlcOpenStatus *status, void *_pParameters, size_t _sParameters, const std::string& taskName  );
 		void subscribe(  const std:: string ModuleName,  const std::string state );
 		bool setTrigger(  );
 		bool setTrue(  );
@@ -45,6 +47,8 @@ namespace atn{
 		unsigned short getPLCOpenState( unsigned short fallback );
         void print( std::ostream &);
 		void plcopenReport( std::ostream &);
+		void printValue( std::ostream &);
 		unsigned int count();
+        unsigned int removeTask( const std::string& taskName );
     };
 }